requirements.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv
firebase-admin
requests
bcrypt
python-multipart